[ADD] refined combine API (#38)
This PR removes the need for the `.publisher(for:)` KVO API call, the `NSObject` inheritance and the `@objc dynamic` annotations by making use of the `projectedValue` property of the property wrappers. See https://www.avanderlee.com/swift/property-wrappers/#projecting-a-value-from-a-property-wrapper for more details on this topic. Please note that I like to use `CurrentValueSubject` instead of `PassthroughSubject` as this makes sure our publisher is stateful and immediately publishes the currently saved value on sink. If we would use `PassthroughSubject` instead, `receiveValue` would only fire after the value has been updated again. 

This is only a draft PR as this makes discussion way easier rather than describing this change in an issue. Also the example view needs  more works and was just thrown together quickly to be able to test the combine API without interference of the manually triggered `tableView.reloadRows` calls after each update of the values. 

Also note the updated deployment targets. This would be required in order to be able to use the Combine publishers

Co-authored-by: Jesse Squires <jesse@jessesquires.com>