Fix crash dereferencing unowned _publisher in KVO trampoline (#112)
A `FoilDefaultStorage{,Optional}` can deallocate while its
`ObserverTrampoline` sticks around long enough to receive a KVO
notification. (I'm not sure how.) When this happens, the attempt to
dereference the captured `_publisher` crashes.

I don't know how to write a reliable test for the test suite, but I can
reproduce the crash ~instantly with this code (e.g. pasted into a new
test) in Xcode 16.0 beta 5 on an iPhone 8 running iOS 16.7.8 (as well as
in iOS 15, 16, 17, and 18 simulators I have kicking around):

```swift
DispatchQueue.global().async {
    var storage: FoilDefaultStorageOptional<Int>?
    while true {
        storage = .init(key: "aloha")
    }
}
while true {
    UserDefaults.standard.set(Int.random(in: Int.min...Int.max), forKey: "aloha")
}
```

That's an admittedly silly test, but I'm seeing crashes in the wild
attempting to dereference `_publisher`.

The obvious fix to me is to capture `_publisher` as either weak or
strong, instead of unowned. A strong reference doesn't cause a cycle, so
I chose it over the theoretically slower weak reference.

---------

Co-authored-by: Jesse Squires <jessesquires@users.noreply.github.com>